pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cpa-lppf-hfl"
version = "0.1.0"
description = "Research artifact for CPA and LPPF in hierarchical federated learning"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Anonymous"}]
dependencies = [
    "torch>=2.1",
    "torchvision>=0.16",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "PyYAML>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[project.scripts]
cpa-lppf = "cpa_lppf.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
